# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "get": {
        "operationId": "Sso_login",
        "parameters": [
          {
            "name": "next",
            "required": false,
            "in": "query",
            "description": "path ในระบบนี้ที่จะกลับไปหลัง login · ไม่ผ่านกฎใช้หน้าแรก",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "ไป {CORE_HUB_WEB_URL}/sso/authorize?subsystem=…&state=… พร้อมคุกกี้ state"
          }
        },
        "security": [],
        "summary": "เริ่มทุกการเข้าสู่ระบบ — สร้าง state แล้วส่งไปเว็บ Core Hub /sso/authorize (ไม่มีหน้าฟอร์ม)",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/callback": {
      "get": {
        "operationId": "Sso_callback",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expires_in",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token_type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "Bearer"
              ],
              "type": "string"
            }
          },
          {
            "name": "access_token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "สำเร็จ — ตั้งคุกกี้ session แล้วไปหน้า next · หรือไม่มี state (เริ่มจาก Core Hub) — ทิ้ง token แล้วไป /auth/login"
          },
          "400": {
            "description": "BAD_REQUEST หรือ VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Central SSO callback — ผลตามข้อ 5.1 ของสัญญา",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/battles": {
      "post": {
        "operationId": "Battles_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBattleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/BattleOutcomeDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST หรือ VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — สถานะปัจจุบันไม่อนุญาต",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "รบ — ท้าทายหอคอย (towerFloor) หรือรบรอบในภูมิภาค (regionRunId)",
        "tags": [
          "battles"
        ]
      },
      "get": {
        "operationId": "Battles_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100000,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "คำค้นหา ไม่สนตัวพิมพ์เล็ก/ใหญ่ · ว่าง = ทั้งหมด",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BattleSummaryDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST หรือ VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "ประวัติการรบแบบย่อ ล่าสุดก่อน · q ค้นในชื่อสถานที่ (หอคอย หรือชื่อภูมิภาค)",
        "tags": [
          "battles"
        ]
      }
    },
    "/api/v1/tower-progress": {
      "get": {
        "operationId": "TowerProgress_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/TowerProgressDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "ความคืบหน้าของหอคอย",
        "tags": [
          "battles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Code Tower — csmju-gamification-knowledge",
    "description": "ระบบย่อย Code Tower ของ CSMJU2030 · ตัวตนมาจาก Core Hub (RS256 + JWKS) · response ทุกตัวห่อด้วย envelope { success, data[, meta] } / { success: false, error }",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "coreHubBearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "ssoCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "core_hub_access_token"
      }
    },
    "schemas": {
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "TOO_MANY_REQUESTS",
              "INTERNAL_ERROR",
              "SERVICE_UNAVAILABLE"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "description": "ใช้กับ VALIDATION_ERROR",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "StatsDto": {
        "type": "object",
        "properties": {
          "str": {
            "type": "number"
          },
          "int": {
            "type": "number"
          },
          "vit": {
            "type": "number"
          },
          "agi": {
            "type": "number"
          },
          "luk": {
            "type": "number"
          }
        },
        "required": [
          "str",
          "int",
          "vit",
          "agi",
          "luk"
        ]
      },
...
      "CharacterDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "classId": {
            "type": "string",
            "enum": [
              "novice",
              "warrior",
              "mage",
              "guardian"
            ]
          },
          "level": {
            "type": "number"
          },
          "exp": {
            "type": "number"
          },
          "expToNext": {
            "type": "number"
          },
          "proficiency": {
            "$ref": "#/components/schemas/ProficiencyDto"
          },
          "stats": {
            "$ref": "#/components/schemas/StatsDto"
          },
          "derived": {
            "description": "ค่าที่คำนวณแล้ว (maxHp, atk, …)",
            "allOf": [
              {
                "$ref": "#/components/schemas/DerivedStatsDto"
              }
            ]
          },
          "gold": {
            "type": "number"
          },
          "materials": {
            "type": "number"
          },
          "equipment": {
            "$ref": "#/components/schemas/EquipmentDto"
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillSummaryDto"
            }
          },
          "highestFloorCleared": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "displayName",
          "classId",
          "level",
          "exp",
          "expToNext",
          "proficiency",
          "stats",
          "derived",
          "gold",
          "materials",
          "equipment",
          "skills",
          "highestFloorCleared"
        ]
      },
...
      "CombatEventDto": {
        "type": "object",
        "properties": {
          "turn": {
            "type": "number"
          },
          "wave": {
            "type": "number"
          },
          "actorId": {
            "type": "string"
          },
          "actorName": {
            "type": "string"
          },
          "action": {
            "type": "string",
            "enum": [
              "attack",
              "skill",
              "defend"
            ]
          },
          "skillId": {
            "type": "string"
          },
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CombatTargetDto"
            }
          },
          "note": {
            "type": "string",
            "enum": [
              "wave_start",
              "wave_clear",
              "defend",
              "windup"
            ]
          },
          "line": {
            "type": "number",
            "description": "บรรทัดในโปรแกรม BloxCode ที่ตัดสินใจเทิร์นนี้ (0 = การกระทำสำรอง)"
          },
          "codeWarnings": {
            "description": "คำเตือนตอนรันโปรแกรม",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mpAfter": {
            "type": "number",
            "description": "MP ของผู้ลงมือหลังจบเทิร์นนี้ (หักค่าร่ายและฟื้นตอนจบเทิร์นแล้ว) — ไม่มีใน event ของระบบ"
          }
        },
        "required": [
          "turn",
          "wave",
          "actorId",
          "actorName",
          "action",
          "targets"
        ]
      },
...
      "EliteDto": {
        "type": "object",
        "properties": {
          "wave": {
            "type": "number"
          },
          "archetypeId": {
            "type": "string"
          },
          "nameTh": {
            "type": "string"
          }
        },
        "required": [
          "wave",
          "archetypeId",
          "nameTh"
        ]
      },
...
      "LevelSyncDto": {
        "type": "object",
        "properties": {
          "level": {
            "type": "number"
          },
          "realLevel": {
            "type": "number"
          },
          "applies": {
            "type": "boolean"
          }
        },
        "required": [
          "level",
          "realLevel",
          "applies"
        ]
      },
      "RegionRunDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "regionId": {
            "type": "string"
          },
          "depth": {
            "type": "number"
          },
          "floor": {
            "type": "number"
          },
          "nameTh": {
            "type": "string"
          },
          "lessonTh": {
            "type": "string"
          },
          "elite": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/EliteDto"
              }
            ]
          },
          "duel": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DuelAnnounceDto"
              }
            ]
          },
          "sync": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LevelSyncDto"
              }
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "regionId",
          "depth",
          "floor",
          "nameTh",
          "lessonTh",
          "expiresAt"
        ]
      },
...
      "BattleDropsDto": {
        "type": "object",
        "properties": {
          "gold": {
            "type": "number"
          },
          "materials": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemDto"
            }
          }
        },
        "required": [
          "gold",
          "materials",
          "items"
        ]
      },
      "BattleResultDto": {
        "type": "object",
        "properties": {
          "victory": {
            "type": "boolean"
          },
          "wavesCleared": {
            "type": "number"
          },
          "events": {
            "description": "บันทึกการรบ (CombatEvent ของ engine)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CombatEventDto"
            }
          },
          "drops": {
            "$ref": "#/components/schemas/BattleDropsDto"
          },
          "expGained": {
            "type": "number"
          },
          "seed": {
            "type": "number"
          }
        },
        "required": [
          "victory",
          "wavesCleared",
          "events",
          "drops",
          "expGained",
          "seed"
        ]
      },
      "DuelBlockDto": {
        "type": "object",
        "properties": {
          "opponent": {
            "$ref": "#/components/schemas/DuelAnnounceDto"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CombatEventDto"
            }
          },
          "won": {
            "type": "boolean"
          },
          "byTimeout": {
            "type": "boolean"
          },
          "rounds": {
            "type": "number"
          },
          "opponentProgram": {
            "type": "string",
            "description": "โปรแกรมของคู่ดวล — ฝ่ายแพ้เปิดดูได้ (รอบ 2W §5.4)"
          }
        },
        "required": [
          "opponent",
          "events",
          "won",
          "byTimeout",
          "rounds",
          "opponentProgram"
        ]
      },
...
      "BattleAttemptDto": {
        "type": "object",
        "properties": {
          "attemptNo": {
            "type": "number",
            "minimum": 1,
            "description": "ครั้งที่ของการรบที่จุดนี้ (1 = ครั้งแรก)"
          },
          "firstAttempt": {
            "type": "boolean",
            "description": "ครั้งแรกที่จุดนี้ — หน้าเว็บให้ดูฉากจนจบก่อน"
          },
          "previous": {
            "nullable": true,
            "description": "ผลครั้งก่อนที่จุดนี้ (null = ไม่เคยรบ)",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PreviousAttemptDto"
              }
            ]
          }
        },
        "required": [
          "attemptNo",
          "firstAttempt",
          "previous"
        ]
      },
      "BattleOutcomeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "id ของบันทึกการรบแบบย่อ (GET /api/v1/battles)"
          },
          "result": {
            "$ref": "#/components/schemas/BattleResultDto"
          },
          "announce": {
            "description": "มีเฉพาะการรบในภูมิภาค",
            "allOf": [
              {
                "$ref": "#/components/schemas/RegionRunDto"
              }
            ]
          },
          "duel": {
            "nullable": true,
            "description": "มีเฉพาะการรบในภูมิภาค",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DuelBlockDto"
              }
            ]
          },
          "character": {
            "description": "ตัวละครหลังบันทึกผลแล้ว",
            "allOf": [
              {
                "$ref": "#/components/schemas/CharacterDto"
              }
            ]
          },
          "leveledUp": {
            "type": "boolean"
          },
          "newLevel": {
            "type": "number"
          },
          "statsGained": {
            "description": "สเตตัสที่ได้จากการเลเวลอัพรอบนี้",
            "allOf": [
              {
                "$ref": "#/components/schemas/StatsDto"
              }
            ]
          },
          "attempt": {
            "$ref": "#/components/schemas/BattleAttemptDto"
          }
        },
        "required": [
          "id",
          "result",
          "character",
          "leveledUp",
          "attempt"
        ]
      },
      "CreateBattleDto": {
        "type": "object",
        "properties": {
          "towerFloor": {
            "type": "number",
            "minimum": 1,
            "maximum": 10,
            "description": "ชั้นของหอคอยที่จะท้าทาย"
          },
          "regionRunId": {
            "type": "string",
            "format": "uuid",
            "description": "id ของรอบจาก POST /api/v1/region-runs"
          }
        }
      },
      "BattleSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "regionId": {
            "type": "string"
          },
          "depth": {
            "type": "number"
          },
          "floor": {
            "type": "number"
          },
          "victory": {
            "type": "boolean"
          },
          "wavesCleared": {
            "type": "number"
          },
          "expGained": {
            "type": "number"
          },
          "goldGained": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "regionId",
          "depth",
          "floor",
          "victory",
          "wavesCleared",
          "expGained",
          "goldGained",
          "createdAt"
        ]
      },
      "TowerProgressDto": {
        "type": "object",
        "properties": {
          "highestFloorCleared": {
            "type": "number",
            "description": "ชั้นสูงสุดของหอคอยที่ผ่านแล้ว (ใบคะแนนของหอคอยเอง)"
          },
          "maxFloor": {
            "type": "number"
          },
          "globalHighestFloor": {
            "type": "number",
            "description": "ความยากสูงสุดที่ผ่านแล้ว \"ที่ไหนก็ได้\" — ตัวปลดล็อกภูมิภาคและไวยากรณ์"
          }
        },
        "required": [
          "highestFloorCleared",
          "maxFloor",
          "globalHighestFloor"
        ]
      },
...
    }
  },
  "security": [
    {
      "coreHubBearer": []
    },
    {
      "ssoCookie": []
    }
  ]
}